general update

+++ src/path_metrics/path_metric_interface.h
@@ -9,7 +9,7 @@ int metric_update(struct path_stats *s, char *server, char *local);
 int populate_path_stats(struct path_stats *ps, char *server, char *local);
 void print_path_stats(struct path_stats *stats, char *id);
 
-int metric_cmp(struct path_stats *ps1, struct path_stats *ps2);
+double metric_cmp(struct path_stats *ps1, struct path_stats *ps2);
 
 /*Memory*/
 struct path_stats *metric_alloc(void);
